attendance-system/requirements.txt
flask==3.1.0
opencv-python==4.10.0.84
face-recognition==1.3.0
numpy==1.26.4
Pillow==10.4.0
